docs(Form): mention of `nuxt-auto-form` for `AutoForm` functionality
There is an increasing demand for generating custom Nuxt UI forms from
schemas. While projects like `shadcn-vue` and `formkit` offer similar
features, they do not integrate directly with Nuxt UI.

To address this issue, I created a `nuxt-auto-form`, module providing
`AutoForm` functionality until it is maybe natively supported in Nuxt UI
(hopefully in v4).

This commit adds a short note to the Form docs, that briefly mentions
the module, which can be a valuable resource for the community.

Related issues: #3876, #4653