Update locale-select.md
The nuxt-i18n docs state that you are not supposed to directly change the value of locale. However, this was the case in the code example, because locale was passed to v-model which caused undesirable behavior. Specifically, this changed the locale before it was loaded from the server. This would result in a console error and the translation key rendering instead of the translation for a brief moment.